+++ src/Masterly.NonEmptyList/NonEmptyList.cs
@@ -1,4 +1,6 @@
-﻿namespace Masterly.NonEmptyList;
+﻿// Ignore Spelling: Deconstruct
+
+namespace Masterly.NonEmptyList;
 
 public class NonEmptyList<T> : List<T>
 {
@@ -28,22 +30,32 @@ public class NonEmptyList<T> : List<T>
 
         Add(firstItem);
 
-        if (otherItems is not null && otherItems.Any())
-            AddRange(otherItems);
+        if (otherItems is not null)
+        {
+            if (otherItems.Any(item => item is null))
+                throw new ArgumentNullException(nameof(otherItems), "Other items cannot contain null values");
+
+            if (otherItems.Any())
+                AddRange(otherItems);
+        }
     }
 
     public new void Add(T item)
     {
-        if (item is null)
-            throw new ArgumentNullException(nameof(item), "Item cannot be null");
+        ArgumentNullException.ThrowIfNull(item);
 
         base.Add(item);
     }
 
     public new void AddRange(IEnumerable<T> collection)
     {
-        if (collection is null || !collection.Any())
-            throw new ArgumentNullException(nameof(collection), "Collection cannot be null");
+        ArgumentNullException.ThrowIfNull(collection);
+
+        if (!collection.Any())
+            throw new ArgumentException("Collection cannot be empty", nameof(collection));
+
+        if (collection.Any(item => item is null))
+            throw new ArgumentNullException(nameof(collection), "Collection cannot contain null values");
 
         base.AddRange(collection);
     }
@@ -60,4 +72,38 @@ public class NonEmptyList<T> : List<T>
     }
 
     public override string ToString() => $"NonEmptyList: [{string.Join(", ", this)}]";
+
+    public new void Clear() => throw new NotSupportedException("Cannot clear a NonEmptyList");
+
+    public new void Remove(T item)
+    {
+        if (Count == 1)
+            throw new InvalidOperationException("Cannot use Remove method while the NonEmptyList contains only one item.");
+
+        base.Remove(item);
+    }
+
+    public new void RemoveAt(int index)
+    {
+        if (Count == 1)
+            throw new InvalidOperationException("Cannot use RemoveAt method while the NonEmptyList contains only one item.");
+
+        base.RemoveAt(index);
+    }
+
+    public new void RemoveAll(Predicate<T> match) => throw new NotSupportedException("RemoveAll method is not supported!");
+
+    public new void RemoveRange(int index, int count)
+    {
+        if (index is 0 && Count == count)
+            throw new InvalidOperationException("Cannot remove all items from a NonEmptyList");
+
+        base.RemoveRange(index, count);
+    }
+
+    public void Deconstruct(out T head, out NonEmptyList<T>? tail)
+    {
+        head = Head;
+        tail = Tail;
+    }
 }